refactor : replace Resp interface to models
