+++ web/templates/gen-windws.html
@@ -21,7 +21,7 @@
                 <div class="input-group">
                     <span class="input-group-text"><i class="fas fa-folder"></i></span>
                     <div class="form-floating">
-                      <input type="text" class="form-control" id="gen-windows-path" name="path" value={{ .tmpPath }} placeholder="디렉토리 경로">
+                      <input type="text" class="form-control" id="gen-windows-path" name="path" value={{ .TmpPath }} placeholder="디렉토리 경로">
                       <label for="gen-windows-path">디렉토리 경로</label>
                     </div>
                 </div>
