+++ websrc/controllers/generateHandlers.go
@@ -7,17 +7,12 @@ import (
 	"path/filepath"
 	"time"
 
+	"github.com/cloud-barista/cm-data-mold/websrc/models"
 	"github.com/labstack/echo/v4"
 )
 
-// SimpleMsg is struct for JSON Simple message
-type BasicResponse struct {
-	Result string `json:"Result"`
-	Error  string `json:"Error"`
-}
-
 type GenerateLinuxPostHandlerResponseBody struct {
-	BasicResponse
+	models.BasicResponse
 }
 
 // GenerateLinuxPostHandler godoc
@@ -39,25 +34,25 @@ func GenerateLinuxPostHandler(ctx echo.Context) error {
 	logger, logstrings := pageLogInit("genlinux", "Create dummy data in linux", start)
 
 	if !osCheck(logger, start, "linux") {
-		return ctx.JSON(http.StatusBadRequest, map[string]interface{}{
-			"Result": logstrings.String(),
-			"Error":  nil,
+		return ctx.JSON(http.StatusBadRequest, models.BasicResponse{
+			Result: logstrings.String(),
+			Error:  nil,
 		})
 	}
 
 	params := getData("gen", ctx).(GenDataParams)
 
 	if !dummyCreate(logger, start, params) {
-		return ctx.JSON(http.StatusInternalServerError, map[string]interface{}{
-			"Result": logstrings.String(),
-			"Error":  nil,
+		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
+			Result: logstrings.String(),
+			Error:  nil,
 		})
 	}
 
 	jobEnd(logger, "Successfully creating a dummy with Linux", start)
-	return ctx.JSON(http.StatusOK, map[string]interface{}{
-		"Result": logstrings.String(),
-		"Error":  nil,
+	return ctx.JSON(http.StatusOK, models.BasicResponse{
+		Result: logstrings.String(),
+		Error:  nil,
 	})
 }
 
@@ -69,9 +64,9 @@ func GenerateWindowsPostHandler(ctx echo.Context) error {
 
 	if !osCheck(logger, start, "windows") {
 		fmt.Println("test")
-		return ctx.JSON(http.StatusBadRequest, map[string]interface{}{
-			"Result": logstrings.String(),
-			"Error":  nil,
+		return ctx.JSON(http.StatusBadRequest, models.BasicResponse{
+			Result: logstrings.String(),
+			Error:  nil,
 		})
 
 	}
@@ -79,17 +74,17 @@ func GenerateWindowsPostHandler(ctx echo.Context) error {
 	params := getData("gen", ctx).(GenDataParams)
 
 	if !dummyCreate(logger, start, params) {
-		return ctx.JSON(http.StatusInternalServerError, map[string]interface{}{
-			"Result": logstrings.String(),
-			"Error":  nil,
+		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
+			Result: logstrings.String(),
+			Error:  nil,
 		})
 
 	}
 
 	jobEnd(logger, "Successfully creating a dummy with Windows", start)
-	return ctx.JSON(http.StatusOK, map[string]interface{}{
-		"Result": logstrings.String(),
-		"Error":  nil,
+	return ctx.JSON(http.StatusOK, models.BasicResponse{
+		Result: logstrings.String(),
+		Error:  nil,
 	})
 }
 
@@ -102,9 +97,9 @@ func GenerateS3PostHandler(ctx echo.Context) error {
 
 	tmpDir, ok := createDummyTemp(logger, start)
 	if !ok {
-		return ctx.JSON(http.StatusInternalServerError, map[string]interface{}{
-			"Result": logstrings.String(),
-			"Error":  nil,
+		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
+			Result: logstrings.String(),
+			Error:  nil,
 		})
 
 	}
@@ -112,34 +107,34 @@ func GenerateS3PostHandler(ctx echo.Context) error {
 	params.DummyPath = tmpDir
 
 	if !dummyCreate(logger, start, params) {
-		return ctx.JSON(http.StatusInternalServerError, map[string]interface{}{
-			"Result": logstrings.String(),
-			"Error":  nil,
+		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
+			Result: logstrings.String(),
+			Error:  nil,
 		})
 
 	}
 
 	awsOSC := getS3OSC(logger, start, "gen", params)
 	if awsOSC == nil {
-		return ctx.JSON(http.StatusInternalServerError, map[string]interface{}{
-			"Result": logstrings.String(),
-			"Error":  nil,
+		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
+			Result: logstrings.String(),
+			Error:  nil,
 		})
 
 	}
 
 	if !oscImport(logger, start, "s3", awsOSC, params.DummyPath) {
-		return ctx.JSON(http.StatusInternalServerError, map[string]interface{}{
-			"Result": logstrings.String(),
-			"Error":  nil,
+		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
+			Result: logstrings.String(),
+			Error:  nil,
 		})
 
 	}
 
 	jobEnd(logger, "Dummy creation and import successful with s3", start)
-	return ctx.JSON(http.StatusOK, map[string]interface{}{
-		"Result": logstrings.String(),
-		"Error":  nil,
+	return ctx.JSON(http.StatusOK, models.BasicResponse{
+		Result: logstrings.String(),
+		Error:  nil,
 	})
 }
 
@@ -151,27 +146,27 @@ func GenerateGCPPostHandler(ctx echo.Context) error {
 	params := getData("gen", ctx).(GenDataParams)
 
 	if !getDataWithBind(logger, start, ctx, &params) {
-		return ctx.JSON(http.StatusInternalServerError, map[string]interface{}{
-			"Result": logstrings.String(),
-			"Error":  nil,
+		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
+			Result: logstrings.String(),
+			Error:  nil,
 		})
 
 	}
 
 	credTmpDir, credFileName, ok := gcpCreateCredFile(logger, start, ctx)
 	if !ok {
-		return ctx.JSON(http.StatusInternalServerError, map[string]interface{}{
-			"Result": logstrings.String(),
-			"Error":  nil,
+		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
+			Result: logstrings.String(),
+			Error:  nil,
 		})
 	}
 	defer os.RemoveAll(credTmpDir)
 
 	tmpDir, ok := createDummyTemp(logger, start)
 	if !ok {
-		return ctx.JSON(http.StatusInternalServerError, map[string]interface{}{
-			"Result": logstrings.String(),
-			"Error":  nil,
+		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
+			Result: logstrings.String(),
+			Error:  nil,
 		})
 
 	}
@@ -179,33 +174,33 @@ func GenerateGCPPostHandler(ctx echo.Context) error {
 	params.DummyPath = tmpDir
 
 	if !dummyCreate(logger, start, params) {
-		return ctx.JSON(http.StatusInternalServerError, map[string]interface{}{
-			"Result": logstrings.String(),
-			"Error":  nil,
+		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
+			Result: logstrings.String(),
+			Error:  nil,
 		})
 
 	}
 
 	gcpOSC := getGCPCOSC(logger, start, "gen", params, credFileName)
 	if gcpOSC == nil {
-		return ctx.JSON(http.StatusInternalServerError, map[string]interface{}{
-			"Result": logstrings.String(),
-			"Error":  nil,
+		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
+			Result: logstrings.String(),
+			Error:  nil,
 		})
 
 	}
 
 	if !oscImport(logger, start, "gcp", gcpOSC, params.DummyPath) {
-		return ctx.JSON(http.StatusInternalServerError, map[string]interface{}{
-			"Result": logstrings.String(),
-			"Error":  nil,
+		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
+			Result: logstrings.String(),
+			Error:  nil,
 		})
 	}
 
 	jobEnd(logger, "Dummy creation and import successful with gcp", start)
-	return ctx.JSON(http.StatusOK, map[string]interface{}{
-		"Result": logstrings.String(),
-		"Error":  nil,
+	return ctx.JSON(http.StatusOK, models.BasicResponse{
+		Result: logstrings.String(),
+		Error:  nil,
 	})
 }
 
@@ -218,9 +213,9 @@ func GenerateNCPPostHandler(ctx echo.Context) error {
 
 	tmpDir, ok := createDummyTemp(logger, start)
 	if !ok {
-		return ctx.JSON(http.StatusInternalServerError, map[string]interface{}{
-			"Result": logstrings.String(),
-			"Error":  nil,
+		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
+			Result: logstrings.String(),
+			Error:  nil,
 		})
 
 	}
@@ -229,33 +224,33 @@ func GenerateNCPPostHandler(ctx echo.Context) error {
 	params.DummyPath = tmpDir
 
 	if !dummyCreate(logger, start, params) {
-		return ctx.JSON(http.StatusInternalServerError, map[string]interface{}{
-			"Result": logstrings.String(),
-			"Error":  nil,
+		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
+			Result: logstrings.String(),
+			Error:  nil,
 		})
 
 	}
 
 	ncpOSC := getS3COSC(logger, start, "gen", params)
 	if ncpOSC == nil {
-		return ctx.JSON(http.StatusInternalServerError, map[string]interface{}{
-			"Result": logstrings.String(),
-			"Error":  nil,
+		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
+			Result: logstrings.String(),
+			Error:  nil,
 		})
 
 	}
 
 	if !oscImport(logger, start, "ncp", ncpOSC, params.DummyPath) {
-		return ctx.JSON(http.StatusInternalServerError, map[string]interface{}{
-			"Result": logstrings.String(),
-			"Error":  nil,
+		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
+			Result: logstrings.String(),
+			Error:  nil,
 		})
 	}
 
 	jobEnd(logger, "Create dummy data and import to ncp objectstorage", start)
-	return ctx.JSON(http.StatusOK, map[string]interface{}{
-		"Result": logstrings.String(),
-		"Error":  nil,
+	return ctx.JSON(http.StatusOK, models.BasicResponse{
+		Result: logstrings.String(),
+		Error:  nil,
 	})
 }
 
@@ -268,9 +263,9 @@ func GenerateMySQLPostHandler(ctx echo.Context) error {
 
 	tmpDir, ok := createDummyTemp(logger, start)
 	if !ok {
-		return ctx.JSON(http.StatusInternalServerError, map[string]interface{}{
-			"Result": logstrings.String(),
-			"Error":  nil,
+		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
+			Result: logstrings.String(),
+			Error:  nil,
 		})
 	}
 	defer os.RemoveAll(tmpDir)
@@ -280,27 +275,27 @@ func GenerateMySQLPostHandler(ctx echo.Context) error {
 	params.SizeServerSQL = "5"
 
 	if !dummyCreate(logger, start, params) {
-		return ctx.JSON(http.StatusInternalServerError, map[string]interface{}{
-			"Result": logstrings.String(),
-			"Error":  nil,
+		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
+			Result: logstrings.String(),
+			Error:  nil,
 		})
 
 	}
 
 	rdbc := getMysqlRDBC(logger, start, "gen", params)
 	if rdbc == nil {
-		return ctx.JSON(http.StatusInternalServerError, map[string]interface{}{
-			"Result": logstrings.String(),
-			"Error":  nil,
+		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
+			Result: logstrings.String(),
+			Error:  nil,
 		})
 
 	}
 
 	sqlList := []string{}
 	if !walk(logger, start, &sqlList, params.DummyPath, ".sql") {
-		return ctx.JSON(http.StatusInternalServerError, map[string]interface{}{
-			"Result": logstrings.String(),
-			"Error":  nil,
+		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
+			Result: logstrings.String(),
+			Error:  nil,
 		})
 	}
 
@@ -313,9 +308,9 @@ func GenerateMySQLPostHandler(ctx echo.Context) error {
 			logger.Errorf("os ReadFile failed : %v", err)
 			logger.Infof("end time : %s", end.Format("2006-01-02T15:04:05-07:00"))
 			logger.Infof("Elapsed time : %s", end.Sub(start).String())
-			return ctx.JSON(http.StatusInternalServerError, map[string]interface{}{
-				"Result": logstrings.String(),
-				"Error":  nil,
+			return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
+				Result: logstrings.String(),
+				Error:  nil,
 			})
 
 		}
@@ -326,18 +321,18 @@ func GenerateMySQLPostHandler(ctx echo.Context) error {
 			logger.Errorf("RDBController import failed : %v", err)
 			logger.Infof("end time : %s", end.Format("2006-01-02T15:04:05-07:00"))
 			logger.Infof("Elapsed time : %s", end.Sub(start).String())
-			return ctx.JSON(http.StatusInternalServerError, map[string]interface{}{
-				"Result": logstrings.String(),
-				"Error":  nil,
+			return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
+				Result: logstrings.String(),
+				Error:  nil,
 			})
 		}
 		logger.Infof("sql put success : %s", filepath.Base(sql))
 	}
 
 	jobEnd(logger, "Dummy creation and import successful with mysql", start)
-	return ctx.JSON(http.StatusOK, map[string]interface{}{
-		"Result": logstrings.String(),
-		"Error":  nil,
+	return ctx.JSON(http.StatusOK, models.BasicResponse{
+		Result: logstrings.String(),
+		Error:  nil,
 	})
 }
 
@@ -350,9 +345,9 @@ func GenerateDynamoDBPostHandler(ctx echo.Context) error {
 
 	tmpDir, ok := createDummyTemp(logger, start)
 	if !ok {
-		return ctx.JSON(http.StatusInternalServerError, map[string]interface{}{
-			"Result": logstrings.String(),
-			"Error":  nil,
+		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
+			Result: logstrings.String(),
+			Error:  nil,
 		})
 	}
 	defer os.RemoveAll(tmpDir)
@@ -362,39 +357,39 @@ func GenerateDynamoDBPostHandler(ctx echo.Context) error {
 	params.SizeServerJSON = "1"
 
 	if !dummyCreate(logger, start, params) {
-		return ctx.JSON(http.StatusInternalServerError, map[string]interface{}{
-			"Result": logstrings.String(),
-			"Error":  nil,
+		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
+			Result: logstrings.String(),
+			Error:  nil,
 		})
 	}
 
 	jsonList := []string{}
 	if !walk(logger, start, &jsonList, params.DummyPath, ".json") {
-		return ctx.JSON(http.StatusInternalServerError, map[string]interface{}{
-			"Result": logstrings.String(),
-			"Error":  nil,
+		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
+			Result: logstrings.String(),
+			Error:  nil,
 		})
 	}
 
 	awsNRDB := getDynamoNRDBC(logger, start, "gen", params)
 	if awsNRDB == nil {
-		return ctx.JSON(http.StatusInternalServerError, map[string]interface{}{
-			"Result": logstrings.String(),
-			"Error":  nil,
+		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
+			Result: logstrings.String(),
+			Error:  nil,
 		})
 	}
 
 	if !nrdbPutWorker(logger, start, "DynamoDB", awsNRDB, jsonList) {
-		return ctx.JSON(http.StatusInternalServerError, map[string]interface{}{
-			"Result": logstrings.String(),
-			"Error":  nil,
+		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
+			Result: logstrings.String(),
+			Error:  nil,
 		})
 	}
 
 	jobEnd(logger, "Dummy creation and import successful with dynamoDB", start)
-	return ctx.JSON(http.StatusOK, map[string]interface{}{
-		"Result": logstrings.String(),
-		"Error":  nil,
+	return ctx.JSON(http.StatusOK, models.BasicResponse{
+		Result: logstrings.String(),
+		Error:  nil,
 	})
 }
 
@@ -405,26 +400,26 @@ func GenerateFirestorePostHandler(ctx echo.Context) error {
 
 	params := GenDataParams{}
 	if !getDataWithBind(logger, start, ctx, &params) {
-		return ctx.JSON(http.StatusInternalServerError, map[string]interface{}{
-			"Result": logstrings.String(),
-			"Error":  nil,
+		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
+			Result: logstrings.String(),
+			Error:  nil,
 		})
 	}
 
 	credTmpDir, credFileName, ok := gcpCreateCredFile(logger, start, ctx)
 	if !ok {
-		return ctx.JSON(http.StatusInternalServerError, map[string]interface{}{
-			"Result": logstrings.String(),
-			"Error":  nil,
+		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
+			Result: logstrings.String(),
+			Error:  nil,
 		})
 	}
 	defer os.RemoveAll(credTmpDir)
 
 	tmpDir, ok := createDummyTemp(logger, start)
 	if !ok {
-		return ctx.JSON(http.StatusInternalServerError, map[string]interface{}{
-			"Result": logstrings.String(),
-			"Error":  nil,
+		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
+			Result: logstrings.String(),
+			Error:  nil,
 		})
 	}
 	defer os.RemoveAll(tmpDir)
@@ -434,39 +429,39 @@ func GenerateFirestorePostHandler(ctx echo.Context) error {
 	params.SizeServerJSON = "1"
 
 	if !dummyCreate(logger, start, params) {
-		return ctx.JSON(http.StatusInternalServerError, map[string]interface{}{
-			"Result": logstrings.String(),
-			"Error":  nil,
+		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
+			Result: logstrings.String(),
+			Error:  nil,
 		})
 	}
 
 	jsonList := []string{}
 	if !walk(logger, start, &jsonList, params.DummyPath, ".json") {
-		return ctx.JSON(http.StatusInternalServerError, map[string]interface{}{
-			"Result": logstrings.String(),
-			"Error":  nil,
+		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
+			Result: logstrings.String(),
+			Error:  nil,
 		})
 	}
 
 	gcpNRDB := getFirestoreNRDBC(logger, start, "gen", params, credFileName)
 	if gcpNRDB == nil {
-		return ctx.JSON(http.StatusInternalServerError, map[string]interface{}{
-			"Result": logstrings.String(),
-			"Error":  nil,
+		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
+			Result: logstrings.String(),
+			Error:  nil,
 		})
 	}
 
 	if !nrdbPutWorker(logger, start, "FirestoreDB", gcpNRDB, jsonList) {
-		return ctx.JSON(http.StatusInternalServerError, map[string]interface{}{
-			"Result": logstrings.String(),
-			"Error":  nil,
+		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
+			Result: logstrings.String(),
+			Error:  nil,
 		})
 	}
 
 	jobEnd(logger, "Dummy creation and import successful with firestoreDB", start)
-	return ctx.JSON(http.StatusOK, map[string]interface{}{
-		"Result": logstrings.String(),
-		"Error":  nil,
+	return ctx.JSON(http.StatusOK, models.BasicResponse{
+		Result: logstrings.String(),
+		Error:  nil,
 	})
 }
 
@@ -479,9 +474,9 @@ func GenerateMongoDBPostHandler(ctx echo.Context) error {
 
 	tmpDir, ok := createDummyTemp(logger, start)
 	if !ok {
-		return ctx.JSON(http.StatusInternalServerError, map[string]interface{}{
-			"Result": logstrings.String(),
-			"Error":  nil,
+		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
+			Result: logstrings.String(),
+			Error:  nil,
 		})
 	}
 	defer os.RemoveAll(tmpDir)
@@ -491,39 +486,39 @@ func GenerateMongoDBPostHandler(ctx echo.Context) error {
 	params.SizeServerJSON = "1"
 
 	if !dummyCreate(logger, start, params) {
-		return ctx.JSON(http.StatusInternalServerError, map[string]interface{}{
-			"Result": logstrings.String(),
-			"Error":  nil,
+		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
+			Result: logstrings.String(),
+			Error:  nil,
 		})
 	}
 
 	jsonList := []string{}
 	if !walk(logger, start, &jsonList, params.DummyPath, ".json") {
-		return ctx.JSON(http.StatusInternalServerError, map[string]interface{}{
-			"Result": logstrings.String(),
-			"Error":  nil,
+		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
+			Result: logstrings.String(),
+			Error:  nil,
 		})
 
 	}
 
 	ncpNRDB := getMongoNRDBC(logger, start, "gen", params)
 	if ncpNRDB == nil {
-		return ctx.JSON(http.StatusInternalServerError, map[string]interface{}{
-			"Result": logstrings.String(),
-			"Error":  nil,
+		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
+			Result: logstrings.String(),
+			Error:  nil,
 		})
 	}
 
 	if !nrdbPutWorker(logger, start, "MongoDB", ncpNRDB, jsonList) {
-		return ctx.JSON(http.StatusInternalServerError, map[string]interface{}{
-			"Result": logstrings.String(),
-			"Error":  nil,
+		return ctx.JSON(http.StatusInternalServerError, models.BasicResponse{
+			Result: logstrings.String(),
+			Error:  nil,
 		})
 	}
 
 	jobEnd(logger, "Dummy creation and import successful with mongoDB", start)
-	return ctx.JSON(http.StatusOK, map[string]interface{}{
-		"Result": logstrings.String(),
-		"Error":  nil,
+	return ctx.JSON(http.StatusOK, models.BasicResponse{
+		Result: logstrings.String(),
+		Error:  nil,
 	})
 }
